requirements.txt
requests
beautifulsoup4
streamlit
pillow